+++ astrocats/catalog/model.py
@@ -1,7 +1,0 @@
-"""Class for representing models."""
-
-from . import struct
-
-Model = struct._Model
-MODEL = Model.get_keychain()
-Model._KEYS = MODEL