+++ astrocats/catalog/quantity.py
@@ -1,7 +1,0 @@
-"""Class for representing quantities."""
-
-from . import struct
-
-Quantity = struct._Quantity
-QUANTITY = Quantity.get_keychain()
-Quantity._KEYS = QUANTITY