+++ astrocats/catalog/realization.py
@@ -1,7 +1,0 @@
-"""Class for representing model realizations."""
-
-from . import struct
-
-Realization = struct._Realization
-REALIZATION = Realization.get_keychain()
-Realization._KEYS = REALIZATION