+++ astrocats/catalog/spectrum.py
@@ -1,7 +1,0 @@
-"""Class for representing spectra."""
-
-from . import struct
-
-Spectrum = struct._Spectrum
-SPECTRUM = Spectrum.get_keychain(extendable=True)
-Spectrum._KEYS = SPECTRUM