+++ astrocats/testcat/tasks/cleanup.py
@@ -5,7 +5,7 @@ import warnings
 from decimal import Decimal
 from math import log10, pi, sqrt
 
-from astrocats.catalog.quantity import QUANTITY
+from astrocats.catalog.struct import QUANTITY
 from astrocats.catalog.utils import (get_sig_digits, is_number, pbar,
                                      pretty_num, tprint, uniq_cdl)
 from astropy import units as un