+++ astrocats/testcat/tasks/general_data.py
@@ -8,7 +8,7 @@ from decimal import Decimal
 from glob import glob
 
 from astrocats.catalog.photometry import PHOTOMETRY
-from astrocats.catalog.spectrum import SPECTRUM
+from astrocats.catalog.struct import SPECTRUM
 from astrocats.catalog.utils import jd_to_mjd, pbar_strings
 from astropy.io import fits
 from astropy.time import Time as astrotime